requirements.txt
# Telegram Bot (Bot API — not userbot)
python-telegram-bot[webhooks]==21.*

# AI / LLM
groq>=1.0.0

# RSS Parsing
feedparser==6.*

# Blockchain / On-chain
web3==6.*

# Async HTTP client
aiohttp==3.9.*

# Sync HTTP client (governance text extraction and fallbacks)
requests==2.*

# HTML parsing / text extraction
beautifulsoup4==4.*

# Scheduler
APScheduler==3.10.*

# Database
SQLAlchemy==2.0.*
asyncpg>=0.29
aiosqlite>=0.19.0

# Config & Env
python-dotenv==1.0.*
pyyaml==6.*

# Timezone
pytz

# Testing
pytest>=7.0
pytest-asyncio>=0.21
pytest-cov>=4.0
aioresponses
